добавил исключение в сервис

+++ src/main/java/com/sadykov/online_shop/application/product/ProductServiceImpl.java
@@ -1,6 +1,7 @@
 package com.sadykov.online_shop.application.product;
 
 import com.sadykov.online_shop.domain.product.Product;
+import com.sadykov.online_shop.domain.product.ProductNotFoundException;
 import com.sadykov.online_shop.domain.product.ProductRepository;
 import org.springframework.stereotype.Service;
 
@@ -23,7 +24,7 @@ public class ProductServiceImpl implements ProductService{
     @Override
     public Product getById(Long id) {
         return productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Product not found: " + id));
+                .orElseThrow(() -> new ProductNotFoundException(id));
     }
 
     @Override
@@ -33,6 +34,10 @@ public class ProductServiceImpl implements ProductService{
 
     @Override
     public void delete(Long id) {
+        Product product = getById(id);
+        if ("ACTIVE".equals(product.getStatus())) {
+            throw new IllegalStateException("Cannot delete ACTIVE product");
+        }
         productRepository.deleteById(id);
     }
 }
